+++ src/main/java/org/folio/listener/kafka/KafkaEventListener.java
@@ -2,8 +2,8 @@ package org.folio.listener.kafka;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
-import java.util.Optional;
 
+import org.folio.domain.dto.Item;
 import org.folio.domain.dto.Loan;
 import org.folio.domain.dto.Request;
 import org.folio.domain.dto.RequestsBatchUpdate;
@@ -12,6 +12,7 @@ import org.folio.domain.dto.UserGroup;
 import org.folio.exception.KafkaEventDeserializationException;
 import org.folio.service.ConsortiaService;
 import org.folio.service.KafkaEventHandler;
+import org.folio.service.impl.ItemEventHandler;
 import org.folio.service.impl.LoanEventHandler;
 import org.folio.service.impl.RequestBatchUpdateEventHandler;
 import org.folio.service.impl.RequestEventHandler;
@@ -23,12 +24,15 @@ import org.folio.spring.FolioModuleMetadata;
 import org.folio.spring.integration.XOkapiHeaders;
 import org.folio.spring.scope.FolioExecutionContextSetter;
 import org.folio.spring.service.SystemUserScopedExecutionService;
-import org.folio.support.KafkaEvent;
+import org.folio.support.kafka.DefaultKafkaEvent;
+import org.folio.support.kafka.InventoryKafkaEvent;
+import org.folio.support.kafka.KafkaEvent;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.messaging.handler.annotation.Headers;
 import org.springframework.stereotype.Component;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -39,9 +43,11 @@ import lombok.extern.log4j.Log4j2;
 @Log4j2
 @RequiredArgsConstructor
 public class KafkaEventListener {
-  private static final ObjectMapper objectMapper = new ObjectMapper();
+  private static final ObjectMapper objectMapper = new ObjectMapper()
+    .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
   private final RequestEventHandler requestEventHandler;
   private final LoanEventHandler loanEventHandler;
+  private final ItemEventHandler itemEventHandler;
   private final UserGroupEventHandler userGroupEventHandler;
   private final UserEventHandler userEventHandler;
   private final SystemUserScopedExecutionService systemUserScopedExecutionService;
@@ -54,7 +60,7 @@ public class KafkaEventListener {
     groupId = "${spring.kafka.consumer.group-id}"
   )
   public void handleRequestEvent(String eventString, @Headers Map<String, Object> messageHeaders) {
-    handleEvent(eventString, requestEventHandler, messageHeaders, Request.class);
+    handleEvent(eventString, requestEventHandler, messageHeaders, DefaultKafkaEvent.class, Request.class);
   }
 
   @KafkaListener(
@@ -62,7 +68,15 @@ public class KafkaEventListener {
     groupId = "${spring.kafka.consumer.group-id}"
   )
   public void handleLoanEvent(String eventString, @Headers Map<String, Object> messageHeaders) {
-    handleEvent(eventString, loanEventHandler, messageHeaders, Loan.class);
+    handleEvent(eventString, loanEventHandler, messageHeaders, DefaultKafkaEvent.class, Loan.class);
+  }
+
+  @KafkaListener(
+    topicPattern = "${folio.environment}\\.\\w+\\.inventory\\.item",
+    groupId = "${spring.kafka.consumer.group-id}"
+  )
+  public void handleItemEvent(String eventString, @Headers Map<String, Object> messageHeaders) {
+    handleEvent(eventString, itemEventHandler, messageHeaders, InventoryKafkaEvent.class, Item.class);
   }
 
   @KafkaListener(
@@ -70,7 +84,7 @@ public class KafkaEventListener {
     groupId = "${spring.kafka.consumer.group-id}"
   )
   public void handleRequestBatchUpdateEvent(String eventString, @Headers Map<String, Object> messageHeaders) {
-    handleEvent(eventString, requestBatchEventHandler, messageHeaders, RequestsBatchUpdate.class);
+    handleEvent(eventString, requestBatchEventHandler, messageHeaders, DefaultKafkaEvent.class, RequestsBatchUpdate.class);
   }
 
   @KafkaListener(
@@ -78,7 +92,7 @@ public class KafkaEventListener {
     groupId = "${spring.kafka.consumer.group-id}"
   )
   public void handleUserGroupEvent(String eventString, @Headers Map<String, Object> messageHeaders) {
-    handleEvent(eventString, userGroupEventHandler, messageHeaders, UserGroup.class);
+    handleEvent(eventString, userGroupEventHandler, messageHeaders, DefaultKafkaEvent.class, UserGroup.class);
   }
 
   @KafkaListener(
@@ -86,20 +100,20 @@ public class KafkaEventListener {
     groupId = "${spring.kafka.consumer.group-id}"
   )
   public void handleUserEvent(String eventString, @Headers Map<String, Object> messageHeaders) {
-    handleEvent(eventString, userEventHandler, messageHeaders, User.class);
+    handleEvent(eventString, userEventHandler, messageHeaders, DefaultKafkaEvent.class, User.class);
   }
 
-  private <T> void handleEvent(String eventString, KafkaEventHandler<T> handler,
-    Map<String, Object> messageHeaders, Class<T> payloadType) {
+  private <E, T> void handleEvent(String eventString, KafkaEventHandler<T> handler,
+    Map<String, Object> messageHeaders, Class<E> kafkaEventClass, Class<T> payloadType) {
 
     log.debug("handleEvent:: event: {}", () -> eventString);
-    KafkaEvent<T> event = deserialize(eventString, messageHeaders, payloadType);
+    KafkaEvent<T> event = deserialize(eventString, messageHeaders, kafkaEventClass, payloadType);
     log.info("handleEvent:: event received: {}", event::getId);
 
     FolioExecutionContext context = DefaultFolioExecutionContext.fromMessageHeaders(
       folioModuleMetadata, messageHeaders);
 
-    try (FolioExecutionContextSetter contextSetter = new FolioExecutionContextSetter(context)) {
+    try (FolioExecutionContextSetter ignored = new FolioExecutionContextSetter(context)) {
       String centralTenantId = consortiaService.getCentralTenantId();
       systemUserScopedExecutionService.executeAsyncSystemUserScoped(centralTenantId,
         () -> handler.handle(event));
@@ -109,17 +123,16 @@ public class KafkaEventListener {
     log.info("handleEvent:: event consumed: {}", event::getId);
   }
 
-  private static <T> KafkaEvent<T> deserialize(String eventString, Map<String, Object> messageHeaders,
-    Class<T> dataType) {
+  private static <E, T> KafkaEvent<T> deserialize(String eventString, Map<String, Object> messageHeaders,
+    Class<E> kafkaEventClass, Class<T> dataType) {
 
     try {
-      JavaType eventType = objectMapper.getTypeFactory()
-        .constructParametricType(KafkaEvent.class, dataType);
-      var kafkaEvent = objectMapper.<KafkaEvent<T>>readValue(eventString, eventType);
-      return Optional.ofNullable(getHeaderValue(messageHeaders, XOkapiHeaders.TENANT))
-        .map(kafkaEvent::withTenantIdHeaderValue)
-        .orElseThrow(() -> new KafkaEventDeserializationException(
-          "Failed to get tenant ID from message headers"));
+      JavaType eventType = objectMapper
+        .getTypeFactory()
+        .constructParametricType(kafkaEventClass, dataType);
+      return objectMapper.<KafkaEvent<T>>readValue(eventString, eventType)
+        .withTenantIdHeaderValue(getHeaderValue(messageHeaders, XOkapiHeaders.TENANT))
+        .withUserIdHeaderValue(getHeaderValue(messageHeaders, XOkapiHeaders.USER_ID));
     } catch (JsonProcessingException e) {
       log.error("deserialize:: failed to deserialize event", e);
       throw new KafkaEventDeserializationException(e);
@@ -129,9 +142,11 @@ public class KafkaEventListener {
   private static String getHeaderValue(Map<String, Object> headers, String headerName) {
     log.debug("getHeaderValue:: headers: {}, headerName: {}", () -> headers, () -> headerName);
     var headerValue = headers.get(headerName);
-    var value = headerValue == null
-      ? null
-      : new String((byte[]) headerValue, StandardCharsets.UTF_8);
+    if (headerValue == null) {
+      throw new KafkaEventDeserializationException(
+        String.format("Failed to get [%s] from message headers", headerName));
+    }
+    var value = new String((byte[]) headerValue, StandardCharsets.UTF_8);
     log.info("getHeaderValue:: header {} value is {}", headerName, value);
     return value;
   }