+++ src/main/java/org/folio/service/impl/RequestEventHandler.java
@@ -2,12 +2,14 @@ package org.folio.service.impl;
 
 import static org.folio.domain.dto.Request.EcsRequestPhaseEnum.PRIMARY;
 import static org.folio.domain.dto.Request.EcsRequestPhaseEnum.SECONDARY;
+import static org.folio.domain.dto.Request.RequestTypeEnum.HOLD;
+import static org.folio.domain.dto.Request.StatusEnum.CLOSED_CANCELLED;
 import static org.folio.domain.dto.TransactionStatus.StatusEnum.AWAITING_PICKUP;
 import static org.folio.domain.dto.TransactionStatus.StatusEnum.CANCELLED;
 import static org.folio.domain.dto.TransactionStatus.StatusEnum.ITEM_CHECKED_OUT;
 import static org.folio.domain.dto.TransactionStatus.StatusEnum.OPEN;
 import static org.folio.support.Constants.INTERIM_SERVICE_POINT_ID;
-import static org.folio.support.KafkaEvent.EventType.UPDATED;
+import static org.folio.support.kafka.EventType.UPDATE;
 
 import java.util.Date;
 import java.util.Objects;
@@ -15,11 +17,11 @@ import java.util.Optional;
 import java.util.UUID;
 import java.util.function.Function;
 
+import org.apache.commons.lang3.StringUtils;
 import org.folio.domain.dto.Request;
 import org.folio.domain.dto.Request.EcsRequestPhaseEnum;
 import org.folio.domain.dto.Request.FulfillmentPreferenceEnum;
 import org.folio.domain.dto.ServicePoint;
-import org.folio.domain.dto.TransactionStatus;
 import org.folio.domain.entity.EcsTlrEntity;
 import org.folio.repository.EcsTlrRepository;
 import org.folio.service.CloningService;
@@ -28,7 +30,7 @@ import org.folio.service.KafkaEventHandler;
 import org.folio.service.RequestService;
 import org.folio.service.ServicePointService;
 import org.folio.spring.service.SystemUserScopedExecutionService;
-import org.folio.support.KafkaEvent;
+import org.folio.support.kafka.KafkaEvent;
 import org.springframework.stereotype.Service;
 
 import lombok.AllArgsConstructor;
@@ -38,6 +40,8 @@ import lombok.extern.log4j.Log4j2;
 @Service
 @Log4j2
 public class RequestEventHandler implements KafkaEventHandler<Request> {
+  // the id used by DCB when canceling a request
+  private static final String DCB_CANCELLATION_REASON_ID = "50ed35b2-1397-4e83-a76b-642adf91ca2a";
 
   private final DcbService dcbService;
   private final EcsTlrRepository ecsTlrRepository;
@@ -49,17 +53,17 @@ public class RequestEventHandler implements KafkaEventHandler<Request> {
   @Override
   public void handle(KafkaEvent<Request> event) {
     log.info("handle:: processing request event: {}", event::getId);
-    if (event.getType() == UPDATED) {
+    if (event.getGenericType() == UPDATE) {
       handleRequestUpdateEvent(event);
     } else {
-      log.info("handle:: ignoring event {} of unsupported type: {}", event::getId, event::getType);
+      log.info("handle:: ignoring event {} of unsupported type: {}", event::getId, event::getGenericType);
     }
     log.info("handle:: request event processed: {}", event::getId);
   }
 
   private void handleRequestUpdateEvent(KafkaEvent<Request> event) {
     log.info("handleRequestUpdateEvent:: handling request update event: {}", event::getId);
-    Request updatedRequest = event.getData().getNewVersion();
+    Request updatedRequest = event.getNewVersion();
     if (updatedRequest == null) {
       log.warn("handleRequestUpdateEvent:: event does not contain new version of request");
       return;
@@ -82,7 +86,7 @@ public class RequestEventHandler implements KafkaEventHandler<Request> {
 
   private void handleRequestUpdateEvent(EcsTlrEntity ecsTlr, KafkaEvent<Request> event) {
     log.debug("handleRequestUpdateEvent:: ecsTlr={}", () -> ecsTlr);
-    Request updatedRequest = event.getData().getNewVersion();
+    Request updatedRequest = event.getNewVersion();
 
     if (!requestMatchesEcsTlr(ecsTlr, updatedRequest, event.getTenantIdHeaderValue())) {
       return;
@@ -98,8 +102,8 @@ public class RequestEventHandler implements KafkaEventHandler<Request> {
   private static boolean requestMatchesEcsTlr(EcsTlrEntity ecsTlr, Request updatedRequest,
     String updatedRequestTenant) {
 
-    final EcsRequestPhaseEnum updatedRequestPhase = updatedRequest.getEcsRequestPhase();
-    final UUID updatedRequestId = UUID.fromString(updatedRequest.getId());
+    EcsRequestPhaseEnum updatedRequestPhase = updatedRequest.getEcsRequestPhase();
+    UUID updatedRequestId = UUID.fromString(updatedRequest.getId());
 
     if (updatedRequestPhase == PRIMARY && updatedRequestId.equals(ecsTlr.getPrimaryRequestId())
       && updatedRequestTenant.equals(ecsTlr.getPrimaryRequestTenantId())) {
@@ -122,7 +126,7 @@ public class RequestEventHandler implements KafkaEventHandler<Request> {
   }
 
   private void handleSecondaryRequestUpdate(EcsTlrEntity ecsTlr, KafkaEvent<Request> event) {
-    processItemIdUpdate(ecsTlr, event.getData().getNewVersion());
+    processItemIdUpdate(ecsTlr, event.getNewVersion());
     updateTransactionStatuses(event, ecsTlr);
   }
 
@@ -140,38 +144,32 @@ public class RequestEventHandler implements KafkaEventHandler<Request> {
     log.info("processItemIdUpdate: ECS TLR {} is updated", ecsTlr::getId);
   }
 
-  private static Optional<TransactionStatus.StatusEnum> determineNewTransactionStatus(
-    KafkaEvent<Request> event) {
+  private void updateTransactionStatuses(KafkaEvent<Request> event, EcsTlrEntity ecsTlr) {
 
-    final Request.StatusEnum oldRequestStatus = event.getData().getOldVersion().getStatus();
-    final Request.StatusEnum newRequestStatus = event.getData().getNewVersion().getStatus();
-    log.info("determineNewTransactionStatus:: oldRequestStatus='{}', newRequestStatus='{}'",
+    var oldRequestStatus = event.getOldVersion().getStatus();
+    var newRequestStatus = event.getNewVersion().getStatus();
+    log.info("updateTransactionStatuses:: oldRequestStatus='{}', newRequestStatus='{}'",
       oldRequestStatus, newRequestStatus);
 
     if (newRequestStatus == oldRequestStatus) {
-      log.info("determineNewTransactionStatus:: request status did not change, doing nothing");
-      return Optional.empty();
+      log.info("updateTransactionStatuses:: request status did not change, doing nothing");
+      return;
     }
 
-    var newTransactionStatus = Optional.ofNullable(
-      switch (newRequestStatus) {
-        case OPEN_IN_TRANSIT -> OPEN;
-        case OPEN_AWAITING_PICKUP -> AWAITING_PICKUP;
-        case CLOSED_FILLED -> ITEM_CHECKED_OUT;
-        case CLOSED_CANCELLED -> CANCELLED;
-        default -> null;
-      });
+    var newTransactionStatus = switch (newRequestStatus) {
+      case OPEN_IN_TRANSIT -> OPEN;
+      case OPEN_AWAITING_PICKUP -> AWAITING_PICKUP;
+      case CLOSED_FILLED -> ITEM_CHECKED_OUT;
+      case CLOSED_CANCELLED -> CANCELLED;
+      default -> null;
+    };
+    if (newTransactionStatus == null) {
+      log.info("updateTransactionStatuses:: irrelevant request status change");
+      return;
+    }
 
-    newTransactionStatus.ifPresentOrElse(
-      ts -> log.info("determineNewTransactionStatus:: new transaction status: {}", ts),
-      () -> log.info("determineNewTransactionStatus:: irrelevant request status change"));
-
-    return newTransactionStatus;
-  }
-
-  private void updateTransactionStatuses(KafkaEvent<Request> event, EcsTlrEntity ecsTlr) {
-    determineNewTransactionStatus(event)
-      .ifPresent(newStatus -> dcbService.updateTransactionStatuses(newStatus, ecsTlr));
+    log.info("updateTransactionStatuses:: new transaction status: {}", newTransactionStatus);
+    dcbService.updateTransactionStatuses(newTransactionStatus, ecsTlr);
   }
 
   private void propagatePrimaryRequestChanges(EcsTlrEntity ecsTlr, KafkaEvent<Request> event) {
@@ -194,7 +192,7 @@ public class RequestEventHandler implements KafkaEventHandler<Request> {
     log.info("propagatePrimaryRequestChanges:: propagating changes to request {} in tenant {}",
       targetRequestId, targetRequestTenantId);
 
-    Request primaryRequest = event.getData().getNewVersion();
+    Request primaryRequest = event.getNewVersion();
     Request targetRequest = requestService.getRequestFromStorage(targetRequestId, targetRequestTenantId);
 
     boolean shouldUpdateTargetRequest = false;
@@ -205,10 +203,9 @@ public class RequestEventHandler implements KafkaEventHandler<Request> {
       targetRequest.setRequestExpirationDate(requestExpirationDate);
       shouldUpdateTargetRequest = true;
     }
-    if (INTERIM_SERVICE_POINT_ID.equals(targetRequest.getPickupServicePointId())) {
-      log.info("propagatePrimaryRequestChanges:: request {} has interim service point as pickup " +
-        "service point, no need to update fulfillment preference", targetRequestId);
-    } else {
+    if (!INTERIM_SERVICE_POINT_ID.equals(targetRequest.getPickupServicePointId())) {
+      log.info("propagatePrimaryRequestChanges:: request {} service point is not interim",
+        targetRequestId);
       if (valueIsNotEqual(primaryRequest, targetRequest, Request::getFulfillmentPreference)) {
         FulfillmentPreferenceEnum fulfillmentPreference = primaryRequest.getFulfillmentPreference();
         log.info("propagatePrimaryRequestChanges:: fulfillment preference changed: {}",
@@ -225,6 +222,16 @@ public class RequestEventHandler implements KafkaEventHandler<Request> {
         clonePickupServicePoint(ecsTlr.getPrimaryRequestTenantId(), targetRequestTenantId,
           pickupServicePointId);
       }
+    }
+
+    if (needToCancelHoldTlr(primaryRequest, targetRequest)) {
+      log.info("propagatePrimaryRequestChanges:: canceling the request: {}", targetRequestId);
+      targetRequest.setStatus(CLOSED_CANCELLED);
+      targetRequest.setCancelledDate(new Date());
+      targetRequest.setCancellationAdditionalInformation("Request cancelled by DCB");
+      targetRequest.setCancelledByUserId(event.getUserIdHeaderValue());
+      targetRequest.setCancellationReasonId(DCB_CANCELLATION_REASON_ID);
+      shouldUpdateTargetRequest = true;
     }
 
     if (!shouldUpdateTargetRequest) {
@@ -250,6 +257,16 @@ public class RequestEventHandler implements KafkaEventHandler<Request> {
       primaryRequestTenantId, () -> servicePointService.find(pickupServicePointId));
     executionService.executeSystemUserScoped(targetRequestTenantId,
       () -> servicePointCloningService.clone(pickupServicePoint));
+  }
+
+  private boolean needToCancelHoldTlr(Request primaryRequest, Request targetRequest) {
+    log.info("needToCancelHoldTlr:: primary request level: {}, type: {}, itemId: {}",
+      primaryRequest.getRequestLevel(), primaryRequest.getRequestType(), primaryRequest.getItemId());
+    return primaryRequest.getRequestLevel() == Request.RequestLevelEnum.TITLE
+      && primaryRequest.getRequestType() == HOLD
+      && StringUtils.isEmpty(primaryRequest.getItemId())
+      && primaryRequest.getStatus() == CLOSED_CANCELLED
+      && targetRequest.getStatus() != CLOSED_CANCELLED;
   }
 
   private static <T, V> boolean valueIsNotEqual(T o1, T o2, Function<T, V> valueExtractor) {