+++ src/main/java/org/folio/service/impl/RequestServiceImpl.java
@@ -1,6 +1,7 @@
 package org.folio.service.impl;
 
 import static java.lang.String.format;
+import static org.folio.support.BarcodeUtil.buildNonEmptyBarcode;
 
 import java.util.Collection;
 import java.util.List;
@@ -225,7 +226,7 @@ public class RequestServiceImpl implements RequestService {
 
     var item = getItemFromStorage(itemId, inventoryTenantId);
     var itemStatus = item.getStatus().getName();
-    var circulationItemStatus = defineCirculationItemStatus(itemStatus);
+    var circulationItemStatus = defineCirculationItemStatus(itemStatus, request.getRequestType());
     log.info("createCirculationItem:: item status {}, calculated status: {}",
       itemStatus, circulationItemStatus);
 
@@ -261,7 +262,7 @@ public class RequestServiceImpl implements RequestService {
       .materialTypeId(item.getMaterialTypeId())
       .permanentLoanTypeId(item.getPermanentLoanTypeId())
       .instanceTitle(instance.getTitle())
-      .barcode(item.getBarcode())
+      .barcode(buildNonEmptyBarcode(item.getBarcode(), item.getId()))
       .pickupLocation(pickupLocation)
       .effectiveLocationId(item.getEffectiveLocationId())
       .lendingLibraryCode("TEST_CODE");
@@ -271,9 +272,9 @@ public class RequestServiceImpl implements RequestService {
   }
 
   private CirculationItemStatus.NameEnum defineCirculationItemStatus(
-    ItemStatus.NameEnum itemStatus) {
+    ItemStatus.NameEnum itemStatus, Request.RequestTypeEnum requestType) {
 
-    return itemStatus == ItemStatus.NameEnum.PAGED
+    return itemStatus == ItemStatus.NameEnum.PAGED && requestType == Request.RequestTypeEnum.PAGE
       ? CirculationItemStatus.NameEnum.AVAILABLE
       : CirculationItemStatus.NameEnum.fromValue(itemStatus.getValue());
   }